Refactor code
Requested changes : Now all functions in controllers and then called in routes files, requireAuth is an available middleware for each routes
